refactor backend + add fixes for stability
* while loop for taking user inputs now have error detection
* pandas warnings are now suppressed to improve display
* increase frequency of data streaming
